pyproject.toml
[project]
name = "anixapi"
version = "1.0.0"
requires-python = ">=3.13"
dependencies = [
    "argon2-cffi~=23.1.0",
    "email-validator~=2.3.0",
    "fastapi~=0.116.1",
    "pydantic~=2.11.7",
    "pyjwt~=2.10.1",
    "python-dotenv~=1.1.1",
    "sqlalchemy~=2.0.38",
    "uvicorn~=0.35.0",
]
